refactor(skills/agentmail): wrap official CLI + use SDK end-to-end
The user pointed out that AgentMail ships an official Python SDK
(`agentmail` on PyPI) and an official npm CLI (`agentmail-cli`). Our
hand-rolled httpx wrapper duplicated work the official tooling already
does. Refactored to lean on upstream:

- Renamed Python package agentmail -> agentmail_bridge to free the
  `agentmail` namespace for the SDK import (the CLI binary stays as
  `agentmail` via the entry point).
- Drop api.py entirely. Replace all REST calls with SDK calls:
  client.agent.sign_up, client.agent.verify, client.inboxes.create,
  client.inboxes.delete, client.webhooks.create, client.webhooks.delete.
- Drop send.py — sending is delegated to the official CLI via passthrough.
- main.py grows a passthrough handler: known Vesta verbs (setup, serve,
  status, teardown) handled in Python; everything else execs the npm
  CLI installed locally at ~/.agentmail/node_modules/.bin/agentmail.
- setup.py now does `npm install --prefix ~/.agentmail agentmail-cli`
  before the webhook step, so a partial setup still leaves the agent
  able to use `agentmail inboxes:messages send` etc.
- pyproject.toml adds `agentmail>=0.4` and renames the project to
  `agentmail-bridge`.

Result: the agent only sees one binary on PATH (`agentmail`). Vesta
verbs are Vesta-specific. All routine ops go through upstream tooling
that auto-updates without us touching code.

Verified end-to-end against live APIs:
- autonomous setup completes through verify (signup, OTP, verify)
- npm install of agentmail-cli succeeds
- passthrough: `agentmail inboxes` returns the upstream CLI help

SKILL.md and SETUP.md rewritten around the wrapper architecture; no
more documentation of our own send/list/etc. — agent uses upstream
verbs.
