+++ agent/skills/agentmail/SETUP.md
@@ -1,12 +1,14 @@
 # agentmail setup
 
-Fully autonomous on the happy path: `agentmail setup` creates a disposable
-mail.tm inbox, signs up to AgentMail with it, polls for the OTP, verifies,
-and discards the disposable inbox. ~2 minutes; no email, no OTP, no clicks.
+Fully autonomous on the happy path: `agentmail setup` provisions an
+AgentMail account, inbox, and webhook end-to-end, plus installs the
+official AgentMail CLI for passthrough, all without asking the user for
+anything. ~2 minutes.
 
-No prerequisites. AgentMail's free tier (3 inboxes, 3,000 sends/month,
-webhook inbound) covers basic agent send + receive without a Cloudflare
-account, a domain, or DNS records.
+No prerequisites beyond `node`/`npm` (used to install the official CLI
+locally). AgentMail's free tier (3 inboxes, 3,000 sends/month, webhook
+inbound) covers basic agent send + receive without a Cloudflare account,
+a domain, or DNS records.
 
 ## 1. Run setup
 
@@ -16,16 +18,20 @@ agentmail setup
 
 What happens, in order:
 
-1. **Disposable inbox** created at mail.tm (no signup required by them).
-2. **AgentMail sign-up** at `POST /v0/agent/sign-up` using the disposable
-   email. Returns `api_key` + `inbox_id` immediately and auto-creates an
-   inbox at `${username}@agentmail.to` (default username:
-   `$AGENT_NAME` lowercased).
+1. **Disposable inbox** created at mail.tm (no signup required).
+2. **AgentMail sign-up** via the official Python SDK
+   (`client.agent.sign_up`). Returns `api_key` + `inbox_id` immediately
+   and auto-creates an inbox at `${username}@agentmail.to`. Default
+   username: `$AGENT_NAME` lowercased.
 3. **OTP poll** on the disposable inbox until AgentMail's verification
-   email arrives (timeout: 3 min); a 6-digit code is extracted via regex.
-4. **Verify** at `POST /v0/agent/verify` with `Bearer ${api_key}` and
-   `{otp_code}`. Persists `AGENTMAIL_API_KEY` to `~/.bashrc`.
-5. **Webhook registration** at `POST /v0/webhooks` pointing at
+   email arrives (timeout: 3 min). A 6-digit code is extracted via regex.
+4. **Verify** via `client.agent.verify` (Bearer-authed with the api_key).
+   Persists `AGENTMAIL_API_KEY` to `~/.bashrc`.
+5. **Install official CLI** locally to `~/.agentmail/node_modules/`
+   (`npm install --prefix ~/.agentmail agentmail-cli`). Idempotent. Done
+   before the webhook step so partial failures still leave the
+   passthrough usable.
+6. **Webhook registration** via `client.webhooks.create` pointing at
    `${VESTAD_TUNNEL}/agents/${AGENT_NAME}/agentmail/webhook?secret=…`.
    Generates `AGENTMAIL_WEBHOOK_SECRET`, persists to `~/.bashrc`.
 
@@ -40,8 +46,8 @@ The autonomous flow can break in two ways:
   no OTP arrives; setup times out at step 3.
 - **mail.tm is down or rate-limiting.** Step 1 fails immediately.
 
-Both hard-exit with a clear message naming the likely cause. Two
-recovery paths:
+Both hard-exit with a clear message naming the likely cause. Two recovery
+paths:
 
 ```bash
 # Manual: prompts for your email and the OTP you receive
@@ -52,7 +58,7 @@ export AGENTMAIL_API_KEY=<paste-key>
 agentmail setup --skip-signup
 ```
 
-## 2. Register and start the local service
+## 2. Register and start the local webhook receiver
 
 The webhook reaches the local FastAPI service through the public vestad
 tunnel; that's why the service must be registered with `"public": true`.
@@ -75,16 +81,19 @@ service comes back up after a container restart.
 
 ## 3. Send a test email
 
+The send command is provided by the official AgentMail CLI, transparently
+passed through by our wrapper:
+
 ```bash
-agentmail send \
+INBOX_ID=$(agentmail status | python3 -c "import sys,json; print(json.load(sys.stdin)['inbox_id'])")
+agentmail inboxes:messages send \
+  --inbox-id "$INBOX_ID" \
   --to <your-personal-email> \
   --subject "test from $AGENT_NAME" \
-  --body "hello"
+  --text "hello"
 ```
 
-A successful send returns `{"ok": true, "result": {"message_id": "<...>", ...}}`.
-Check the recipient inbox to confirm delivery, then reply from there and
-watch for the inbound notification:
+Reply from your personal inbox, then watch for the inbound notification:
 
 ```bash
 ls -la ~/agent/notifications/ | grep agentmail
@@ -111,7 +120,7 @@ re-registration needed.
 agentmail teardown
 ```
 
-Deletes the inbox via the AgentMail API and clears
-`~/.agentmail/config.json`. The AgentMail account itself is left alone (an
-account can hold multiple inboxes; deleting the account requires the
-console).
+Deletes the inbox + webhook via the AgentMail API and clears
+`~/.agentmail/config.json`. The AgentMail account itself and the locally
+installed npm CLI are left in place (an account can hold multiple
+inboxes; deleting the account requires the console).
