+++ agent/skills/agentmail/cli/pyproject.toml
@@ -1,9 +1,10 @@
 [project]
-name = "agentmail"
+name = "agentmail-bridge"
 version = "0.1.0"
-description = "CLI for the agentmail skill"
+description = "Vesta-side bridge to the official AgentMail CLI/SDK"
 requires-python = ">=3.11"
 dependencies = [
+    "agentmail>=0.4",
     "fastapi>=0.115",
     "uvicorn>=0.30",
     "httpx>=0.27",
@@ -11,11 +12,11 @@ dependencies = [
 ]
 
 [project.scripts]
-agentmail = "agentmail.main:main"
+agentmail = "agentmail_bridge.main:main"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/agentmail"]
+packages = ["src/agentmail_bridge"]
