+++ agent/skills/agentmail/cli/src/agentmail_bridge/disposable_mail.py
